+++ src/Sensors/Loom_Analog/Loom_Analog.cpp
@@ -4,7 +4,7 @@
 void Loom_Analog::measure() {
 
     // Read the data from the given analog pin
-    for (int i = 0; i < pinMappings.size(); i++) {
+    for (size_t i = 0; i < pinMappings.size(); i++) {
 
         /* If we are measuring the Vbat pin we want a little different behavior */
         if (pinMappings[i]->pinNumber == A7) {
@@ -60,7 +60,7 @@ float Loom_Analog::analogToMV(int analog) {
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////
 float Loom_Analog::getMV(int pin) {
-    for (int i = 0; i < pinMappings.size(); i++) {
+    for (size_t i = 0; i < pinMappings.size(); i++) {
         if (pinMappings[i]->pinNumber == pin) {
             return pinMappings[i]->analog_mv;
         }
@@ -71,7 +71,7 @@ float Loom_Analog::getMV(int pin) {
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////
 float Loom_Analog::getAnalog(int pin) {
-    for (int i = 0; i < pinMappings.size(); i++) {
+    for (size_t i = 0; i < pinMappings.size(); i++) {
         if (pinMappings[i]->pinNumber == pin) {
             return pinMappings[i]->analog;
         }
